+++ src/compiler.c
@@ -56,18 +56,18 @@ struct instruction *compiler_current_instructions(struct compiler *c) {
     return scope.instructions;
 }
 
-unsigned int 
+size_t 
 add_instruction(struct compiler *c, struct instruction *ins) {
     struct instruction *cins = compiler_current_instructions(c);
-    unsigned int pos = cins->size;
+    size_t pos = cins->size;
 
-    unsigned int new_size = cins->size + ins->size;
+    size_t new_size = cins->size + ins->size;
     if (new_size >= cins->cap) {
         cins->cap *= 2;
         cins->bytes = realloc(cins->bytes, cins->cap * sizeof(*cins->bytes));
     }
 
-    for (int i=0; i < ins->size; i++) {
+    for (size_t i=0; i < ins->size; i++) {
         cins->bytes[cins->size++] = ins->bytes[i];
     }
 
@@ -75,7 +75,7 @@ add_instruction(struct compiler *c, struct instruction *ins) {
     return pos;
 }
 
-unsigned int 
+size_t
 add_constant(struct compiler *c, struct object *obj) {
     // TODO: Dereference here?
     c->constants->values[c->constants->size++] = obj;
@@ -99,7 +99,7 @@ void compiler_remove_last_instruction(struct compiler *c) {
 }
 
 void compiler_replace_instruction(struct compiler *c, unsigned int pos, struct instruction *ins) {
-    for (int i=0; i < ins->size; i++) {
+    for (size_t i=0; i < ins->size; i++) {
         c->scopes[c->scope_index].instructions->bytes[pos + i] = ins->bytes[i];
     }
 
@@ -107,7 +107,7 @@ void compiler_replace_instruction(struct compiler *c, unsigned int pos, struct i
 }
 
 void compiler_replace_last_instruction(struct compiler *c, struct instruction *ins) {
-    int pos = compiler_current_scope(c).last_instruction.position;
+    size_t pos = compiler_current_scope(c).last_instruction.position;
     compiler_replace_instruction(c, pos, ins);
     compiler_set_last_instruction(c, ins->bytes[0], pos);
 }
@@ -120,18 +120,18 @@ bool compiler_last_instruction_is(struct compiler *c, enum opcode opcode) {
     return c->scopes[c->scope_index].last_instruction.opcode == opcode;
 }
 
-void compiler_change_operand(struct compiler *c, unsigned int pos, int operand) {
+void compiler_change_operand(struct compiler *c, size_t pos, int operand) {
     enum opcode opcode = c->scopes[c->scope_index].instructions->bytes[pos];
     struct instruction *new = make_instruction(opcode, operand);
     compiler_replace_instruction(c, pos, new);
 }
 
-unsigned int compiler_emit(struct compiler *c, enum opcode opcode, ...) {
+size_t compiler_emit(struct compiler *c, enum opcode opcode, ...) {
     va_list args;
     va_start(args, opcode);
     struct instruction *ins = make_instruction_va(opcode, args);
     va_end(args);
-    unsigned int pos = add_instruction(c, ins);
+    size_t pos = add_instruction(c, ins);
     compiler_set_last_instruction(c, opcode, pos);
     return pos;
 }
@@ -139,7 +139,7 @@ unsigned int compiler_emit(struct compiler *c, enum opcode opcode, ...) {
 int
 compile_program(struct compiler *compiler, struct program *program) {
     int err;
-    for (int i=0; i < program->size; i++) {
+    for (size_t i=0; i < program->size; i++) {
         err = compile_statement(compiler, &program->statements[i]);
         if (err) return err;
     }
@@ -150,7 +150,7 @@ compile_program(struct compiler *compiler, struct program *program) {
 int
 compile_block_statement(struct compiler *compiler, struct block_statement *block) {
     int err;
-    for (int i=0; i < block->size; i++) {
+    for (size_t i=0; i < block->size; i++) {
         err = compile_statement(compiler, &block->statements[i]);
         if (err) return err;
     }
@@ -266,7 +266,7 @@ compile_expression(struct compiler *c, struct expression *expr) {
             if (err) return err;
 
             /* we don't know where to jump yet, so we use 9999 as placeholder */
-            unsigned int jump_if_not_true_pos = compiler_emit(c, OPCODE_JUMP_NOT_TRUE, 9999);
+            size_t jump_if_not_true_pos = compiler_emit(c, OPCODE_JUMP_NOT_TRUE, 9999);
 
             err = compile_block_statement(c, expr->ifelse.consequence);
             if (err) { return err; }
@@ -275,8 +275,8 @@ compile_expression(struct compiler *c, struct expression *expr) {
                 compiler_remove_last_instruction(c);
             }
 
-            unsigned int jump_pos = compiler_emit(c, OPCODE_JUMP, 9999);
-            unsigned int after_conseq_pos = c->scopes[c->scope_index].instructions->size;
+            size_t jump_pos = compiler_emit(c, OPCODE_JUMP, 9999);
+            size_t after_conseq_pos = c->scopes[c->scope_index].instructions->size;
             compiler_change_operand(c, jump_if_not_true_pos, after_conseq_pos);
 
             if (expr->ifelse.alternative) {
@@ -290,7 +290,7 @@ compile_expression(struct compiler *c, struct expression *expr) {
                 compiler_emit(c, OPCODE_NULL);
             }
 
-            unsigned int after_alternative_pos = c->scopes[c->scope_index].instructions->size;
+            size_t after_alternative_pos = c->scopes[c->scope_index].instructions->size;
             compiler_change_operand(c, jump_pos, after_alternative_pos);
         }
         break;
@@ -327,7 +327,7 @@ compile_expression(struct compiler *c, struct expression *expr) {
         case EXPR_FUNCTION: {
             compiler_enter_scope(c);
 
-            for (int i=0; i < expr->function.parameters.size; i++) {
+            for (size_t i=0; i < expr->function.parameters.size; i++) {
                 symbol_table_define(c->symbol_table, expr->function.parameters.values[i].value);
             }
 
@@ -340,7 +340,7 @@ compile_expression(struct compiler *c, struct expression *expr) {
                 compiler_emit(c, OPCODE_RETURN);
             }
 
-            unsigned int num_locals = c->symbol_table->size;
+            size_t num_locals = c->symbol_table->size;
             struct instruction *ins = compiler_leave_scope(c);
             struct object *obj = make_compiled_function_object(ins, num_locals);
             compiler_emit(c, OPCODE_CONST, add_constant(c, obj));
@@ -351,7 +351,7 @@ compile_expression(struct compiler *c, struct expression *expr) {
             err = compile_expression(c, expr->call.function);
             if (err) return err;
 
-            int i = 0;
+            size_t i = 0;
             for (; i < expr->call.arguments.size; i++) {
                 err = compile_expression(c, expr->call.arguments.values[i]);
                 if (err) return err;
@@ -373,6 +373,7 @@ struct bytecode *
 get_bytecode(struct compiler *c) {
     struct bytecode *b = malloc(sizeof *b);
     b->instructions = compiler_current_instructions(c);
+    // TODO: Copy object list here?
     b->constants = c->constants;
     return b;
 }
